Solution to Input Two Floats and Print Their Product

+++ Input Two Floats and Print Their Product@@6788d5012db25efec946f93e/code.c
@@ -4,7 +4,7 @@ int main(){
     float num1;
     float num2;
     scanf("%f %f",&num1,&num2);
-    int product = num1*num2;
+    float product = num1*num2;
     printf("Product: %f",product);
 
 }
